Space Station & Movie Theatre
Space Station, this page is an example home page for a dog walking services. It's styled using CSS and contains a navbar, image containers and a form.  Movie Theater, this page is an example home page for a Movie Theater that is styled using Bootstrap 4. It contains many Bootstrap components including but not limited to a navbar, jumbotron, form, cards and dropdowns.